requirements.txt
appdirs==1.4.4
attrs==21.2.0
Authlib==0.15.5
beautifulsoup4==4.10.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
cryptography==3.4.8
decorator==5.1.0
defusedxml==0.7.1
dnspython==2.1.0
docutils==0.18.1
et-xmlfile==1.1.0
exchangelib==4.6.2
fpdf2==2.4.6
furl==2.1.3
graphviz==0.13.2
htmldocx==0.0.6
idna==3.3
isodate==0.6.1
java-access-bridge-wrapper==0.7.5
jsonpath-ng==1.5.3
jsonschema==4.3.1
lxml==4.7.1
mss==6.1.0
netsuitesdk==1.24.0
notifiers==1.2.1
ntlm-auth==1.5.0
oauthlib==3.1.1
openpyxl==3.0.9
orderedmultidict==1.0.1
pdfminer.six==20201018
Pillow==8.4.0
platformdirs==2.4.0
ply==3.11
psutil==5.8.0
pycparser==2.21
Pygments==2.10.0
pynput-robocorp-fork==4.0.0
PyPDF2==1.26.0
pyperclip==1.8.2
pyrsistent==0.18.0
PySocks==1.7.1
python-docx==0.8.11
pytz==2021.3
pywin32==301
pywinauto==0.6.8
PyYAML==5.4.1
requests==2.26.0
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3987==1.3.8
robocorp-dialog==0.4.1
robotframework==4.1.3
robotframework-pythonlibcore==3.0.0
robotframework-requests==0.9.2
robotframework-sapguilibrary==1.1
robotframework-seleniumlibrary==5.1.3
robotframework-seleniumtestability==1.1.0
rpaframework==12.2.0
rpaframework-core==6.5.1
rpaframework-dialogs==0.4.2
rpaframework-pdf==1.30.4
selenium==3.141.0
simple-salesforce==1.11.4
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tenacity==8.0.1
tqdm==4.62.3
tweepy==3.10.0
tzdata==2021.5
tzlocal==2.1
urllib3==1.26.7
webdrivermanager==0.10.0
wrapt==1.13.3
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.1.0
